Rename section SASS folder to pages
