+++ src/assets/scss/pages/Contact.scss
@@ -1,7 +1,7 @@
 @use "../utils/" as *;
 
 .contact {
-  &__section {
+  &__page {
     background-color: var(--bg-200);
 
     .container {
