+++ src/assets/scss/pages/Home.scss
@@ -1,7 +1,7 @@
 @use "../utils/" as *;
 
 .home {
-  &__section {
+  &__page {
     background-color: var(--bg-100);
 
     .title {
