+++ src/assets/scss/pages/about/_page.scss
@@ -1,7 +1,7 @@
 @use "../../utils/" as *;
 
 .about {
-  &__section {
+  &__page {
     background-color: var(--bg-200);
 
     section {
