+++ src/assets/scss/pages/about/_soft-skills.scss
