+++ src/assets/scss/pages/about/_tech.scss
