+++ src/assets/scss/pages/about/index.scss
@@ -1,4 +1,4 @@
-@forward "section";
+@forward "page";
 @forward "bio";
 @forward "soft-skills";
 @forward "tech";
