+++ src/assets/scss/pages/portfolio/_page.scss
@@ -12,7 +12,7 @@
 
   &__project {
     &--list {
-      grid-template-columns: repeat(auto-fit, minmax(25rem, 1fr));
+      grid-template-columns: repeat(auto-fit, minmax(23rem, 1fr));
       gap: var(--spacing, 1rem);
     }
