+++ src/assets/scss/pages/portfolio/index.scss
@@ -1,3 +1,3 @@
-@forward "section";
+@forward "page";
 @forward "card";
 @forward "modal";
